// API_AutoUpdateDB/enums.ts
/**
 * Enum type data
 */
export enum TypeData{
    customer = 'customer',
    product = 'product',
    order = 'order',

    // 2023.04.10
    temp_design = 'temp_design',
}

/**
 * Enum type update
 */
 export enum TypeUpdate{
    create = 'create',
    update = 'update',
    delete = 'delete',
    cancelled = 'cancelled',
    fulfilled = 'fulfilled',
    paid = 'paid',
    partially_fulfilled = 'partially_fulfilled',
}